Fixed HTTP OPTIONS issue
There was a problem with client browsers connecting from http://localhost to the server even if Access-Control-Allow-Origin "*" was already set in .htaccess. The issue
was that the PHP Slim framework (correctly) recognizes OPTIONS as a valid verb and attempts to proceed with authorization of API key. Therefore, proper handling of the
OPTIONS verb needed to be handled in the Slim middleware and api.php modules. Currenty we simply return empty body as rfc2616 does not state that body is required.
HTTP headers are already adjusted in the .htaccess file.