Added cache headers to HTTP respons (issue #12 )
Entity tag (ETag) headers are now added for the following resources:
/additives, /additives/{code}, /categories/{code}. The ETag unique-id
value is calculated using the 'last_update' SQL schema property for each
item selected. Results will be cached in Redis and the ETag unique-id will
be recalculated on every request. The server will return (304) Not
Modified for matching ETag ids.

This is not a full fix for the issue because results are still fetched and
sent to the user agent. A proper fix would be to directly compare the ETag
ids without actually making a Redis data fetch request!