Fixed Slim Stop Exception issue (issue #12)
As soon as the Slim framework detects that user agent sent ETag matches
the server produced ETag set in the Model classes it halts the execution
with (304) Not Modified and throws a \Slim\Exception\Stop exception which
was being handled by the try-catch (Exception) in Model class methods.
This in turn was then rethrown as ModelException and that cause incorrect
HTTP (500) Internal Server Error to be thrown.

This whole scenario actually happens when there was NO cache at all on
server side (in Redis) but the user agent (browser) still sends the ETag
with the request.

Now fixed by catching the Stop exception in Model class methods and
rethrowing it to the caller.